# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NivenX Hosting API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  email: string;
  password: string;
  name: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  stripeCustomerId?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Plan {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  description?: string | null;
  priceMonthly: number;
  priceYearly: number;
  /** @nullable */
  stripePriceIdMonthly?: string | null;
  /** @nullable */
  stripePriceIdYearly?: string | null;
  features: string[];
  /** @nullable */
  ram?: string | null;
  /** @nullable */
  cpu?: string | null;
  /** @nullable */
  storage?: string | null;
  /** @nullable */
  bandwidth?: string | null;
  isPopular: boolean;
  isActive: boolean;
}

export interface PlanInput {
  name: string;
  category: string;
  description?: string;
  priceMonthly: number;
  priceYearly: number;
  stripePriceIdMonthly?: string;
  stripePriceIdYearly?: string;
  features: string[];
  ram?: string;
  cpu?: string;
  storage?: string;
  bandwidth?: string;
  isPopular?: boolean;
  isActive?: boolean;
}

/**
 * @nullable
 */
export type PriceRecurring = { [key: string]: unknown } | null;

export interface Price {
  id: string;
  /** @nullable */
  unit_amount: number | null;
  currency: string;
  /** @nullable */
  recurring?: PriceRecurring;
  active?: boolean;
}

export interface ProductWithPrices {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  active?: boolean;
  prices: Price[];
}

export interface ProductsWithPricesResponse {
  data: ProductWithPrices[];
}

export interface CheckoutInput {
  priceId: string;
}

export interface CheckoutResponse {
  url: string;
}

export interface PortalResponse {
  url: string;
}

/**
 * @nullable
 */
export type SubscriptionResponseSubscription = { [key: string]: unknown } | null;

export interface SubscriptionResponse {
  /** @nullable */
  subscription?: SubscriptionResponseSubscription;
}

export interface Service {
  id: number;
  userId: number;
  planId: number;
  name: string;
  status: string;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  cpuUsage?: number | null;
  /** @nullable */
  ramUsage?: number | null;
  /** @nullable */
  diskUsage?: number | null;
  /** @nullable */
  bandwidthUsed?: number | null;
  createdAt: string;
}

export interface ServiceInput {
  planId: number;
  name: string;
  location: string;
}

export interface ServiceActionInput {
  action: string;
}

export interface ServiceActionResult {
  success: boolean;
  message: string;
}

export interface Invoice {
  id: number;
  userId: number;
  /** @nullable */
  stripeInvoiceId?: string | null;
  amount: number;
  currency?: string;
  status: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface Ticket {
  id: number;
  userId: number;
  subject: string;
  status: string;
  priority: string;
  createdAt: string;
  updatedAt?: string;
}

export interface TicketInput {
  subject: string;
  message: string;
  priority: string;
  department?: string;
}

export interface TicketReply {
  id: number;
  ticketId: number;
  userId: number;
  message: string;
  isStaff?: boolean;
  createdAt: string;
}

export interface TicketDetail {
  id: number;
  userId: number;
  subject: string;
  status: string;
  priority: string;
  createdAt: string;
  updatedAt?: string;
  replies: TicketReply[];
}

export interface TicketReplyInput {
  message: string;
}

export interface Announcement {
  id: number;
  title: string;
  content: string;
  pinned?: boolean;
  createdAt: string;
}

export interface AnnouncementInput {
  title: string;
  content: string;
  pinned?: boolean;
}

export interface KbArticle {
  id: number;
  title: string;
  content: string;
  category: string;
  slug?: string;
  createdAt: string;
}

export interface DashboardStats {
  activeServices: number;
  openTickets: number;
  pendingInvoices: number;
  totalSpent: number;
}

export interface AdminStats {
  totalUsers: number;
  totalRevenue: number;
  activeServices: number;
  openTickets: number;
  totalInvoices: number;
}

export interface SiteSettings {
  id: number;
  siteName: string;
  /** @nullable */
  tagline: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  faviconUrl?: string | null;
  primaryColor?: string;
  accentColor?: string;
  maintenanceMode: boolean;
  /** @nullable */
  discordUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
  /** @nullable */
  youtubeUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  tiktokUrl?: string | null;
}

export interface SiteSettingsInput {
  siteName?: string;
  tagline?: string;
  primaryColor?: string;
  accentColor?: string;
  maintenanceMode?: boolean;
  discordUrl?: string;
  twitterUrl?: string;
  youtubeUrl?: string;
  instagramUrl?: string;
  tiktokUrl?: string;
}

export interface AdminUserUpdate {
  role?: string;
  name?: string;
  email?: string;
}