# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NivenX Hosting API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # Plans
...
  /plans/{id}:
    get:
      operationId: getPlan
      tags: [plans]
      summary: Get a plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "404":
          description: Not found

  # Stripe products/prices
...
  /stripe/checkout:
    post:
      operationId: createCheckout
      tags: [billing]
      summary: Create Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
...
  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/plans:
    get:
      operationId: listAdminPlans
      tags: [admin]
      summary: List plans (admin)
      responses:
        "200":
          description: Plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
    post:
      operationId: createAdminPlan
      tags: [admin]
      summary: Create plan (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /admin/plans/{id}:
    patch:
      operationId: updateAdminPlan
      tags: [admin]
      summary: Update plan (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    delete:
      operationId: deleteAdminPlan
      tags: [admin]
      summary: Delete plan (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /admin/announcements:
    post:
      operationId: createAdminAnnouncement
      tags: [admin]
      summary: Create announcement (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get site settings (admin)
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
    patch:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update site settings (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        stripeCustomerId:
          type: ["string", "null"]
        createdAt:
          type: string
    Plan:
      type: object
      required: [id, name, category, priceMonthly, priceYearly, features, isPopular, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        priceMonthly:
          type: number
        priceYearly:
          type: number
        stripePriceIdMonthly:
          type: ["string", "null"]
        stripePriceIdYearly:
          type: ["string", "null"]
        features:
          type: array
          items:
            type: string
        ram:
          type: ["string", "null"]
        cpu:
          type: ["string", "null"]
        storage:
          type: ["string", "null"]
        bandwidth:
          type: ["string", "null"]
        isPopular:
          type: boolean
        isActive:
          type: boolean
    PlanInput:
      type: object
      required: [name, category, priceMonthly, priceYearly, features]
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        priceMonthly:
          type: number
        priceYearly:
          type: number
        stripePriceIdMonthly:
          type: string
        stripePriceIdYearly:
          type: string
        features:
          type: array
          items:
            type: string
        ram:
          type: string
        cpu:
          type: string
        storage:
          type: string
        bandwidth:
          type: string
        isPopular:
          type: boolean
        isActive:
          type: boolean
...
    CheckoutInput:
      type: object
      required: [priceId]
      properties:
        priceId:
          type: string
    CheckoutResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
...
    Announcement:
      type: object
      required: [id, title, content, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        pinned:
          type: boolean
        createdAt:
          type: string
    AnnouncementInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        pinned:
          type: boolean
...
    SiteSettings:
      type: object
      required: [id, siteName, tagline, maintenanceMode]
      properties:
        id:
          type: integer
        siteName:
          type: string
        tagline:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        faviconUrl:
          type: ["string", "null"]
        primaryColor:
          type: string
        accentColor:
          type: string
        maintenanceMode:
          type: boolean
        discordUrl:
          type: ["string", "null"]
        twitterUrl:
          type: ["string", "null"]
        youtubeUrl:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        tiktokUrl:
          type: ["string", "null"]
    SiteSettingsInput:
      type: object
      properties:
        siteName:
          type: string
        tagline:
          type: string
        primaryColor:
          type: string
        accentColor:
          type: string
        maintenanceMode:
          type: boolean
        discordUrl:
          type: string
        twitterUrl:
          type: string
        youtubeUrl:
          type: string
        instagramUrl:
          type: string
        tiktokUrl:
          type: string
    AdminUserUpdate:
      type: object
      properties:
        role:
          type: string
        name:
          type: string
        email:
          type: string