# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NivenX Hosting API
 * OpenAPI spec version: 0.1.0
 */

export interface Plan {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  description?: string | null;
  priceMonthly: number;
  priceYearly: number;
  /** @nullable */
  stripePriceIdMonthly?: string | null;
  /** @nullable */
  stripePriceIdYearly?: string | null;
  features: string[];
  /** @nullable */
  ram?: string | null;
  /** @nullable */
  cpu?: string | null;
  /** @nullable */
  storage?: string | null;
  /** @nullable */
  bandwidth?: string | null;
  isPopular: boolean;
  isActive: boolean;
}