# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NivenX Hosting API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettings {
  id: number;
  siteName: string;
  /** @nullable */
  tagline: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  faviconUrl?: string | null;
  primaryColor?: string;
  accentColor?: string;
  maintenanceMode: boolean;
  /** @nullable */
  discordUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
  /** @nullable */
  youtubeUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  tiktokUrl?: string | null;
}