# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NivenX Hosting API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  email: string;
  name: string;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  stripeCustomerId?: string | null;
  createdAt: string;
}